requirements.txt
# Core RAG & NLP
langchain==0.1.10
langchain-community==0.0.20
sentence-transformers==2.2.2
chromadb==0.4.24
pypdf==4.0.1

# LLM API
openai==1.3.9
openrouter==0.1.0

# Web Framework
streamlit==1.28.1

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
